#9 Add the ability to use Nomic embedding model

+++ chat.py
@@ -28,6 +28,7 @@ DATABASE_PATH = os.getenv('DATABASE_PATH', '')
 COLLECTION_NAME = os.getenv('COLLECTION_NAME', 'works')
 PERSIST_DIRECTORY = os.getenv('PERSIST_DIRECTORY', 'works_db')
 MODEL = os.getenv('MODEL', 'gpt-4o')
+EMBEDDINGS_MODEL = os.getenv('EMBEDDINGS_MODEL', 'nomic-embed-text')
 LLM_BASE_URL = os.getenv('LLM_BASE_URL', None)
 REBUILD = os.getenv('REBUILD', 'false').lower() == 'true'
 HISTORY = os.getenv('HISTORY', 'true').lower() == 'true'
@@ -41,7 +42,7 @@ if MODEL.startswith('gpt'):
     embeddings = OpenAIEmbeddings()
 else:
     llm = Ollama(model=MODEL)
-    embeddings = OllamaEmbeddings(model=MODEL)
+    embeddings = OllamaEmbeddings(model=EMBEDDINGS_MODEL)
     if LLM_BASE_URL:
         llm.base_url = LLM_BASE_URL
         embeddings.base_url = LLM_BASE_URL
@@ -57,40 +58,47 @@ if len(docsearch) < 1 or REBUILD:
         'results': [],
     }
     params = {'page': ''}
-    if ALL:
-        print('Loading all of the works from the ACMI Public API')
-        while True:
-            page_data = requests.get(
-                'https://api.acmi.net.au/works/',
-                params=params,
-                timeout=10,
-            ).json()
-            json_data['results'].extend(page_data['results'])
-            if not page_data.get('next'):
-                break
-            params['page'] = furl(page_data.get('next')).args.get('page')
-            if len(json_data['results']) % 1000 == 0:
-                print(f'Downloaded {len(json_data["results"])}...')
-    else:
-        print('Loading the first ten pages of works from the ACMI Public API')
-        PAGES = 10
-        json_data = {
-            'results': [],
-        }
-        for index in range(1, (PAGES + 1)):
-            page_data = requests.get(
-                'https://api.acmi.net.au/works/',
-                params=params,
-                timeout=10,
-            )
-            json_data['results'].extend(page_data.json()['results'])
-            print(f'Downloaded {page_data.request.url}')
-            params['page'] = furl(page_data.json().get('next')).args.get('page')
-    print(f'Finished downloading {len(json_data["results"])} works.')
-
     TMP_FILE_PATH = 'data.json'
-    with open(TMP_FILE_PATH, 'w', encoding='utf-8') as json_file:
-        json.dump(json_data, json_file)
+
+    if os.path.isfile(TMP_FILE_PATH):
+        print('Loading works from the ACMI Public API data.json file you have already created...')
+        with open(TMP_FILE_PATH, 'r') as tmp_file:
+            json_data = json.load(tmp_file)
+    else:
+        if ALL:
+            print('Loading all of the works from the ACMI Public API')
+            while True:
+                page_data = requests.get(
+                    'https://api.acmi.net.au/works/',
+                    params=params,
+                    timeout=10,
+                ).json()
+                json_data['results'].extend(page_data['results'])
+                if not page_data.get('next'):
+                    break
+                params['page'] = furl(page_data.get('next')).args.get('page')
+                if len(json_data['results']) % 1000 == 0:
+                    print(f'Downloaded {len(json_data["results"])}...')
+        else:
+            print('Loading the first ten pages of works from the ACMI Public API')
+            PAGES = 10
+            json_data = {
+                'results': [],
+            }
+            for index in range(1, (PAGES + 1)):
+                page_data = requests.get(
+                    'https://api.acmi.net.au/works/',
+                    params=params,
+                    timeout=10,
+                )
+                json_data['results'].extend(page_data.json()['results'])
+                print(f'Downloaded {page_data.request.url}')
+                params['page'] = furl(page_data.json().get('next')).args.get('page')
+        print(f'Finished downloading {len(json_data["results"])} works.')
+
+        with open(TMP_FILE_PATH, 'w', encoding='utf-8') as json_file:
+            json.dump(json_data, json_file)
+
     json_loader = JSONLoader(
         file_path=TMP_FILE_PATH,
         jq_schema='.results[]',
